allow prefix to be added for blueprint removal

+++ RollbackBlueprint.ps1
@@ -1,3 +1,5 @@
+param([string]$Prefix)
+
 $subscriptionId = (az account show --query id --output tsv)
 if ($LastExitCode -ne 0) {
     throw "An error has occured. Subscription id query failed."
@@ -6,9 +8,20 @@ if ($LastExitCode -ne 0) {
 $list = az blueprint assignment list --subscription $subscriptionId | ConvertFrom-Json
 $list | ForEach-Object {
     $name = $_.name
-    az blueprint assignment delete --name $name --subscription $subscriptionId --yes  
-    if ($LastExitCode -ne 0) {
-        throw "An error has occured. Rollback assignment $name failed."
+
+    $shouldContinue = $true
+    if ($Prefix -and !$name.Contains($Prefix)) {
+        $shouldContinue = $false
+    }
+
+    if ($shouldContinue) {
+        az blueprint assignment delete --name $name --subscription $subscriptionId --yes  
+        if ($LastExitCode -ne 0) {
+            throw "An error has occured. Rollback assignment $name failed."
+        }
+    }
+    else {
+        Write-Host "Skipping $name"
     }
 }
 
@@ -19,20 +32,31 @@ if ($LastExitCode -ne 0) {
 
 $blueprints | ForEach-Object {
     $name = $_.name
-    $versions = az blueprint version list --blueprint-name $name --subscription $subscriptionId | ConvertFrom-Json
-    if ($LastExitCode -ne 0) {
-        throw "An error has occured. Listing all versions for $name failed."
-    }
-    $versions | ForEach-Object {
-        $version = $_.name
-        az blueprint version delete --blueprint-name $name --subscription $subscriptionId --version $version --yes
-        if ($LastExitCode -ne 0) {
-            throw "An error has occured. Rollbacking back version $version failed."
-        }
+
+    $shouldContinue = $true
+    if ($Prefix -and !$name.Contains($Prefix)) {
+        $shouldContinue = $false
     }
 
-    az blueprint delete --name $name --subscription $subscriptionId --yes
-    if ($LastExitCode -ne 0) {
-        throw "An error has occured. Removing blueprint $name failed."
+    if ($shouldContinue) {
+        $versions = az blueprint version list --blueprint-name $name --subscription $subscriptionId | ConvertFrom-Json
+        if ($LastExitCode -ne 0) {
+            throw "An error has occured. Listing all versions for $name failed."
+        }
+        $versions | ForEach-Object {
+            $version = $_.name
+            az blueprint version delete --blueprint-name $name --subscription $subscriptionId --version $version --yes
+            if ($LastExitCode -ne 0) {
+                throw "An error has occured. Rollbacking back version $version failed."
+            }
+        }
+
+        az blueprint delete --name $name --subscription $subscriptionId --yes
+        if ($LastExitCode -ne 0) {
+            throw "An error has occured. Removing blueprint $name failed."
+        }
+    }
+    else {
+        Write-Host "Skipping $name"
     }
 }
